Move validation code to single file & remove unneeded functions.
